requirements.txt
certifi==2022.9.24
charset-normalizer==2.0.12
idna==3.4
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.5
requests==2.27.1
six==1.16.0
urllib3==1.26.12
instaloader==4.9.5
setuptools==28.8.0

